deduplicate fixtures
